backend/requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
pylint
psycopg2
pytest-django
black